System.X509.Unix

Description

this module is portable to unix system where there is usually a etcssl/certs with system X509 certificates.

the path can be dynamically override using the environment variable defined by envPathOverride in the module, which by default is SYSTEM_CERTIFICATE_PATH

Documentation

getSystemCertificateStore :: IO CertificateStore